Enhance AI Settings Page and Connection Testing Features
- Updated the `ai-settings.php` file to improve the layout and functionality of the AI provider configuration section, including separate fields for API keys for OpenAI, Anthropic, and Hugging Face.
- Implemented a more user-friendly interface for testing API connections, with loading indicators and clear success/error messages.
- Enhanced the `HuggingFaceAdapter` class to improve the test connection method, adding detailed logging for better debugging and error handling.
- Updated the `KeywordGenerator` class to ensure scores are explicitly cast to float, preventing implicit type conversion issues.
- Improved overall documentation to reflect the new features and enhancements, ensuring adherence to project standards.

+++ wp-content/plugins/asapdigest-core/includes/ai/processors/class-keyword-generator.php
@@ -363,7 +363,7 @@ class KeywordGenerator {
                     if ($count > 1) { // Only include phrases that appear multiple times
                         $phrases[] = [
                             'keyword' => $phrase,
-                            'score' => min(1.0, $count / 10),
+                            'score' => (float)min(1.0, $count / 10), // Cast to float to prevent implicit conversion
                             'occurrences' => $count
                         ];
                     }
